Refactor CommandException and Command
