+++ src/main/java/seedu/address/logic/commands/management/DeleteLessonCommand.java
@@ -2,7 +2,6 @@ package seedu.address.logic.commands.management;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.exceptions.CommandException.MESSAGE_EXPECTED_MGT_MODEL;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.CommandHistory;
