+++ src/main/java/seedu/address/logic/commands/quiz/QuizStatusCommand.java
@@ -1,7 +1,5 @@
 package seedu.address.logic.commands.quiz;
 
-import static seedu.address.logic.commands.exceptions.CommandException.MESSAGE_EXPECTED_QUIZ_MODEL;
-
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
