+++ src/test/java/seedu/address/logic/commands/management/DeleteLessonCommandTest.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands.management;
 import static java.util.Objects.requireNonNull;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
-import static seedu.address.logic.commands.exceptions.CommandException.MESSAGE_EXPECTED_MGT_MODEL;
+import static seedu.address.logic.commands.Command.MESSAGE_EXPECTED_MGT_MODEL;
 import static seedu.address.model.Lessons.EXCEPTION_INVALID_INDEX;
 import static seedu.address.testutil.TypicalLessons.LESSON_DEFAULT;
